Create README.md
The developed information system provides capabilities of searching and retrieving of qualitative and quantitative information of research activities, information which is produced and collected automatically through web crawling techniques in real time. In particular, the system categorizes and manages a continuous flow of raw data in order to produce information that leads to qualitative and quantitative evaluation metrics. It presents both the raw data and the added valuable data and information in an intelligent and effective way adapted to the background and the interests of the viewer using data visualization techniques.